+++ praw/objects.py
@@ -24,6 +24,7 @@ that it can be saved and unsaved in the context of a logged in user.
 
 from __future__ import print_function, unicode_literals
 
+import json
 import six
 from six.moves.urllib.parse import (  # pylint: disable=F0401
     parse_qs, urlparse, urlunparse)
@@ -838,6 +839,15 @@ class Redditor(Gildable, Messageable, Refreshable):
         """
         return self.reddit_session.get_multireddit(self, multi)
 
+    def get_multireddits(self):
+        """Return the list of multireddits belonging to this user.
+
+        If this is the signed in user, all multis are visible.
+        Otherwise, only public multireddits are seen.
+
+        """
+        return self.reddit_session.get_multireddits(self.name)
+
     def mark_as_read(self, messages, unread=False):
         """Mark message(s) as read or unread.
 
@@ -1454,8 +1464,104 @@ class Multireddit(Refreshable):
                                                                  multi)
         super(Multireddit, self).__init__(reddit_session, json_dict, fetch,
                                           info_url)
+        self.name = multi
+        self.update_urls(reddit_session, redditor, multi)
+
+    def __repr__(self):
+        """Return a code representation of the Multireddit."""
+        return 'Multireddit(author=\'{0}\', name=\'{1}\')'.format(
+            self.author, self._name)
+
+    @restrict_access(scope='subscribe')
+    def add_subreddit(self, subreddit, delete=False, *args, **kwargs):
+        """Add a subreddit to the user's multireddit.
+
+        :param subreddit: The subreddit name or Subreddit object to add
+        :param delete: Optionally, convert this request into a DELETE,
+            removing the subreddit from the multireddit.
+        Any additional parameters are passed directly into
+        :meth:`request_json` of the reddit_session
+
+        """
+        subreddit = six.text_type(subreddit)
+        url = self.reddit_session.config['multireddit_add']
+        url = url % (self.author.name, self.name, subreddit)
+        multipath = 'user/%s/m/%s/' % (self.author.name, self.name)
+
+        data = {
+            'multipath': multipath,
+            'srname': subreddit}
+        if delete:
+            method = 'DELETE'
+        else:
+            method = 'PUT'
+            model = json.dumps({'name': subreddit})
+            data['model'] = model
+        return self.reddit_session.request_json(url, data=data, method=method,
+                                                *args, **kwargs)
+
+    @restrict_access(scope='subscribe')
+    def copy(self, to_name):
+        """Copy this multireddit.
+
+        This function is a handy shortcut to
+        :meth:`copy_multireddit` of the reddit_session
+
+        """
+        return self.reddit_session.copy_multireddit(self.author, self.name,
+                                                    to_name)
+
+    @restrict_access(scope='subscribe')
+    def edit(self, description=None, icon_name=None,
+             key_color=None, subreddits=None, visibility=None,
+             weighting_scheme=None,
+             *args, **kwargs):
+        """Edit this multireddit.
+
+        This function is a handy shortcut to
+        :meth:`edit_multireddit` of the reddit_session
+
+        """
+        return self.reddit_session.edit_multireddit(
+            name=self.name,
+            description=description,
+            icon_name=icon_name,
+            key_color=key_color,
+            subreddits=subreddits,
+            visibility=visibility,
+            weighting_scheme=weighting_scheme,
+            *args, **kwargs)
+
+    @restrict_access(scope='subscribe')
+    def remove_subreddit(self, subreddit, *args, **kwargs):
+        """Remove a subreddit from the user's multireddit."""
+        return self.add_subreddit(subreddit, True, *args, **kwargs)
+
+    @restrict_access(scope='subscribe')
+    def rename(self, new_name, *args, **kwargs):
+        """Rename this multireddit.
+
+        This function is a handy shortcut to
+        :meth:`rename_multireddit` of the reddit_session.
+
+        """
+        response = self.reddit_session.rename_multireddit(self.name, new_name,
+                                                          *args, **kwargs)
+        self.name = new_name
+        # self._name = new_name
+        self.update_urls(self.reddit_session, self.author.name, new_name)
+        return response
+
+    def update_urls(self, reddit_session, redditor, multi):
+        """Rebuild URLs for this object.
+
+        This is necessary because calling :meth:`rename` breaks
+        all sorters and paths.
+
+        """
         self._name = multi
         self.author = redditor
+        self.path = '/user/%s/m/%s' % (redditor, multi)
         self._url = reddit_session.config['multireddit'] % (self.author,
                                                             self._name)
 
@@ -1464,12 +1570,8 @@ class Multireddit(Refreshable):
                                                         self._name))
         self._listing_urls = [base + x + '.json' for x in listings]
         self.subreddits = [
-            Subreddit(reddit_session, x['name']) for x in self.subreddits]
-
-    def __repr__(self):
-        """Return a code representation of the Multireddit."""
-        return 'Multireddit(author=\'{0}\', name=\'{1}\')'.format(
-            self.author, self._name)
+            Subreddit(reddit_session, x['name']) for x in self.subreddits
+            if isinstance(x, dict)]
 
     def __unicode__(self):
         """Return a string representation of the Multireddit."""